requirements.txt
openai >= 1.0
huggingface_hub
python-dotenv
tenacity
matplotlib
transformers
torch
bitsandbytes
jupyter
accelerate
protobuf
sentencepiece
scipy
datasets